# Improve performance of access to coin in TxOut