--- a/eras/babbage/impl/src/Cardano/Ledger/Babbage/Collateral.hs
+++ b/eras/babbage/impl/src/Cardano/Ledger/Babbage/Collateral.hs
@@ -6,20 +6,24 @@
 
 -- | Figure 2: Functions related to fees and collateral
 --   Babbage Specification
-module Cardano.Ledger.Babbage.Collateral where
+module Cardano.Ledger.Babbage.Collateral
+  ( isTwoPhaseScriptAddress,
+    collAdaBalance,
+    collOuts,
+  )
+where
 
 import Cardano.Ledger.Address (Addr (..))
 import Cardano.Ledger.Alonzo.Tx (isTwoPhaseScriptAddressFromMap)
-import Cardano.Ledger.Alonzo.TxBody (AlonzoEraTxBody (..))
 import Cardano.Ledger.Alonzo.TxInfo (ExtendedUTxO (txscripts))
 import Cardano.Ledger.Babbage.TxBody (BabbageEraTxBody (..), BabbageTxBody, BabbageTxOut (..))
 import Cardano.Ledger.BaseTypes (TxIx (..), txIxFromIntegral)
 import Cardano.Ledger.Block (txid)
+import Cardano.Ledger.Coin (Coin)
 import Cardano.Ledger.Core
-import Cardano.Ledger.Shelley.UTxO (UTxO (..), balance)
+import Cardano.Ledger.Shelley.UTxO (UTxO (..), coinBalance)
 import Cardano.Ledger.TxIn (TxIn (..))
 import Cardano.Ledger.Val ((<->))
-import Control.SetAlgebra (eval, (◁))
 import qualified Data.Map.Strict as Map
 import Data.Maybe.Strict (StrictMaybe (..))
 import Data.Word (Word16, Word64)
@@ -36,19 +40,17 @@ isTwoPhaseScriptAddress ::
   Bool
 isTwoPhaseScriptAddress tx utxo = isTwoPhaseScriptAddressFromMap @era (txscripts utxo tx)
 
-collBalance ::
-  forall era.
+collAdaBalance ::
   (BabbageEraTxBody era, TxOut era ~ BabbageTxOut era) =>
   TxBody era ->
-  UTxO era ->
-  Value era
-collBalance txBody (UTxO m) =
+  Map.Map (TxIn (Crypto era)) (TxOut era) ->
+  Coin
+collAdaBalance txBody utxoCollateral =
   case txBody ^. collateralReturnTxBodyL of
     SNothing -> colbal
-    SJust txOut -> colbal <-> (txOut ^. valueTxOutL @era)
+    SJust txOut -> colbal <-> (txOut ^. coinTxOutL)
   where
-    col = UTxO (eval ((txBody ^. collateralInputsTxBodyL) ◁ m))
-    colbal = balance @era col
+    colbal = coinBalance $ UTxO utxoCollateral
 
 collOuts ::
   ( BabbageEraTxBody era,