… eras/alonzo/test-suite/src/Test/Cardano/Ledger/Alonzo/AlonzoEraGen.hs
… eras/shelley/test-suite/src/Test/Cardano/Ledger/Shelley/Generator/Utxo.hs
… eras/shelley/test-suite/src/Test/Cardano/Ledger/Shelley/Rules/TestChain.hs
… libs/cardano-ledger-core/src/Cardano/Ledger/Core.hs
… libs/cardano-ledger-test/src/Test/Cardano/Ledger/Examples/BabbageFeatures.hs
… libs/cardano-ledger-test/src/Test/Cardano/Ledger/Generic/Functions.hs
… libs/ledger-state/README.md
… libs/ledger-state/bench/Performance.hs