Performance of `areAllAdaOnly` and `coinBalance` are over 100x faster than the one of `balance` (see benchmark below). This is because whenever we only need `Coin` we don't have to deserialize the `MaryValue` and we can operate on the compact form of the coin directly.

Therefore this PR attempts to switch all places where this pattern happens

* `coin $ balance utxo` -> `coinBalance utxo`
* `isAdaOnly $ balance utxo` -> `areAllAdaOnly utxo`
* `coin (txOut ^. valueTxOutL)` -> `txOut ^. coinTxOutL`

Another big part of this PR is the explicit implementation of the fact that it is possible to have MultiAsset in the collateral input set, but only if it is fully being send back in the returnCollateral. Most importantly it changes the predicate failure `CollateralContainsNonADA` to not return negative MultiAsset values, when returnCollateral contains mismatched type and/or amount of concrete MultiAsset.

---
These are the benchmarks of the newly added functions on the full UTxO, except `areAllAdaOnly`, which is applied to the part of the UTxO that is actually all ada only.
```
benchmarking UTxO/balance
time                 15.16 s    (13.32 s .. 16.53 s)
                     0.998 R²   (0.994 R² .. 1.000 R²)
mean                 14.82 s    (14.31 s .. 15.14 s)
std dev              507.2 ms   (195.3 ms .. 675.9 ms)
variance introduced by outliers: 19% (moderately inflated)

benchmarking UTxO/coinBalance
time                 147.4 ms   (146.8 ms .. 147.7 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 147.2 ms   (146.9 ms .. 147.4 ms)
std dev              317.9 μs   (177.1 μs .. 449.6 μs)
variance introduced by outliers: 12% (moderately inflated)

benchmarking UTxO/areAllAdaOnly
time                 114.6 ms   (111.9 ms .. 115.9 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 115.6 ms   (114.7 ms .. 116.2 ms)
std dev              1.003 ms   (529.9 μs .. 1.439 ms)
variance introduced by outliers: 11% (moderately inflated)
```